Create README.md
The Robot Grid World Simulator is in two parts: A Simulator (RobotGridWorld.pyw) which simulates a simple grid square based world, and also a Client robot control programmers library (grobot.py), which enables the creation and control of one or more robots within the simple grid world simulator. The grobot.py library API allows to create and control a simulated robot in the simulated GridWorld. 

The robot can navigate through the maze and explore as many of the unoccupied grid squares of the GridWorld as possible. Also